Closes #4, Login page
