+++ lib/data/data.md
@@ -1,0 +1,7 @@
+### Readme
+
+`Data` folder represents `Store`.
+
+I'm naming this `Data` instead of `Store` because the Materiel Icon Theme converts Data folder with style and ignores a folder named Store.
+
+:wink:
